flask-backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.24.0
opencv-python>=4.5.3
scikit-learn>=1.0.0
pillow>=8.3.2
joblib>=1.0.1
python-dotenv>=0.19.0
werkzeug==2.0.1